Refactor: divide in smaller component
