+++ src/components/SiteFooter/SiteCopyright.jsx
@@ -1,0 +1,9 @@
+function SiteCopyright() {
+  return (
+    <p className="text-(--grey) md:order-2 md:justify-self-end">
+      © 2021 Loopstudios. All rights reserved.
+    </p>
+  );
+}
+
+export default SiteCopyright;
