+++ src/components/SiteFooter/SiteFooterNavigation.jsx
@@ -1,0 +1,15 @@
+function SiteFooterNavigation() {
+  return (
+    <nav className="py-4 md:order-1 md:justify-self-start">
+      <ul className="links">
+        <li>About</li>
+        <li>Careers</li>
+        <li>Events</li>
+        <li>Products</li>
+        <li>Support</li>
+      </ul>
+    </nav>
+  );
+}
+
+export default SiteFooterNavigation;
