+++ src/components/SiteFooter/SiteLogo.jsx
@@ -1,0 +1,5 @@
+function SiteLogo() {
+  return <img src="/images/logo.svg" className="md:justify-self-start " />;
+}
+
+export default SiteLogo;
